Rearrange test contents to match the new source-file layout.
Following the source split, move each function's tests into the test file named
after its source:
- `isFabricValue()` -> new `type-check.test.ts`
- `tagFromNativeValue()` / `tagFromNativeClass()` -> new `native-type-tags.test.ts`
- conversion functions (`shallowFabricFromNativeValue()`,
  `fabricFromNativeValue()`, `nativeFromFabricValue()`, `isFabricCompatible()`)
  -> `native-conversion.test.ts`; the round-trip tests are merged into its
  existing `nativeFromFabricValue()` describe.
- `fabric-value.test.ts` slimmed to `valueEqual()` (with a new test for it).

Also normalize function-named `describe()`s to `name()` form per conventions,
and prune the now-unused `native-type-tags` import.
