+++ packages/data-model/test/native-type-tags.test.ts
@@ -1,0 +1,177 @@
+import { describe, it } from "@std/testing/bdd";
+import { expect } from "@std/expect";
+import {
+  NATIVE_TAGS,
+  tagFromNativeClass,
+  tagFromNativeValue,
+} from "../src/native-type-tags.ts";
+
+describe("native-type-tags", () => {
+  describe("tagFromNativeValue()", () => {
+    it("returns `Error` tag for standard `Error` subclasses", () => {
+      const cases: [string, Error][] = [
+        ["Error", new Error("test")],
+        ["TypeError", new TypeError("test")],
+        ["RangeError", new RangeError("test")],
+        ["SyntaxError", new SyntaxError("test")],
+        ["ReferenceError", new ReferenceError("test")],
+        ["URIError", new URIError("test")],
+        ["EvalError", new EvalError("test")],
+      ];
+      for (const [_name, value] of cases) {
+        expect(tagFromNativeValue(value)).toBe(NATIVE_TAGS.Error);
+      }
+    });
+
+    it("returns `Error` tag for exotic `Error` subclass (custom class)", () => {
+      class MyFancyError extends Error {
+        constructor(msg: string) {
+          super(msg);
+          this.name = "MyFancyError";
+        }
+      }
+      const exotic = new MyFancyError("exotic");
+      // Constructor is MyFancyError, not in the switch -- falls back to
+      // Error.isError().
+      expect(tagFromNativeValue(exotic)).toBe(NATIVE_TAGS.Error);
+    });
+
+    it("returns `Map` tag for `Map` instances", () => {
+      expect(tagFromNativeValue(new Map())).toBe(NATIVE_TAGS.Map);
+    });
+
+    it("returns `Set` tag for `Set` instances", () => {
+      expect(tagFromNativeValue(new Set())).toBe(NATIVE_TAGS.Set);
+    });
+
+    it("returns `Date` tag for `Date` instances", () => {
+      expect(tagFromNativeValue(new Date())).toBe(NATIVE_TAGS.Date);
+    });
+
+    it("returns `Uint8Array` tag for `Uint8Array` instances", () => {
+      expect(tagFromNativeValue(new Uint8Array())).toBe(
+        NATIVE_TAGS.Uint8Array,
+      );
+    });
+
+    it("returns `Object` tag for plain objects", () => {
+      expect(tagFromNativeValue({})).toBe(NATIVE_TAGS.Object);
+    });
+
+    it("returns `Array` tag for arrays", () => {
+      expect(tagFromNativeValue([])).toBe(NATIVE_TAGS.Array);
+    });
+
+    it("returns `RegExp` tag for `RegExp` instances", () => {
+      expect(tagFromNativeValue(/abc/)).toBe(NATIVE_TAGS.RegExp);
+    });
+
+    it("returns `Object` tag for null-prototype objects (no constructor)", () => {
+      const obj = Object.create(null);
+      expect(tagFromNativeValue(obj)).toBe(NATIVE_TAGS.Object);
+    });
+
+    it("returns `HasToJSON` tag for plain objects with `toJSON()`", () => {
+      const obj = { toJSON: () => "converted" };
+      expect(tagFromNativeValue(obj)).toBe(NATIVE_TAGS.HasToJSON);
+    });
+
+    it("returns `HasToJSON` tag for arrays with `toJSON()`", () => {
+      const arr = [1, 2, 3] as unknown[] & { toJSON?: () => unknown };
+      arr.toJSON = () => "custom array";
+      expect(tagFromNativeValue(arr)).toBe(NATIVE_TAGS.HasToJSON);
+    });
+
+    it("returns `HasToJSON` tag for class instances with `toJSON()`", () => {
+      class Custom {
+        toJSON() {
+          return { x: 1 };
+        }
+      }
+      expect(tagFromNativeValue(new Custom())).toBe(NATIVE_TAGS.HasToJSON);
+    });
+
+    it("returns `Date` tag for `Date` (not `HasToJSON` despite `Date.toJSON`)", () => {
+      expect(tagFromNativeValue(new Date())).toBe(NATIVE_TAGS.Date);
+    });
+
+    // Functions are non-objects and return Primitive from tagFromNativeValue.
+    // In practice, functions with toJSON() are handled separately in
+    // the modern conversion path, not via tagFromNativeValue.
+    it("returns `Primitive` for functions (even with `toJSON`)", () => {
+      const fn = () => {};
+      (fn as unknown as { toJSON: () => string }).toJSON = () => "converted";
+      expect(tagFromNativeValue(fn)).toBe(NATIVE_TAGS.Primitive);
+    });
+  });
+
+  describe("tagFromNativeClass()", () => {
+    it("returns `Error` tag for standard `Error` constructors", () => {
+      const constructors = [
+        Error,
+        TypeError,
+        RangeError,
+        SyntaxError,
+        ReferenceError,
+        URIError,
+        EvalError,
+      ];
+      for (const ctor of constructors) {
+        expect(tagFromNativeClass(ctor)).toBe(NATIVE_TAGS.Error);
+      }
+    });
+
+    it("returns `Error` tag for exotic `Error` subclass constructor", () => {
+      class ExoticError extends Error {}
+      // Constructor is ExoticError, not in the switch -- falls back to
+      // Error.isError(prototype) check.
+      expect(tagFromNativeClass(ExoticError)).toBe(NATIVE_TAGS.Error);
+    });
+
+    it("returns correct tags for `Array`, `Object`, `Map`, `Set`, `Date`, `Uint8Array`", () => {
+      expect(tagFromNativeClass(Array)).toBe(NATIVE_TAGS.Array);
+      expect(tagFromNativeClass(Object)).toBe(NATIVE_TAGS.Object);
+      expect(tagFromNativeClass(Map)).toBe(NATIVE_TAGS.Map);
+      expect(tagFromNativeClass(Set)).toBe(NATIVE_TAGS.Set);
+      expect(tagFromNativeClass(Date)).toBe(NATIVE_TAGS.Date);
+      expect(tagFromNativeClass(Uint8Array)).toBe(NATIVE_TAGS.Uint8Array);
+    });
+
+    it("returns `RegExp` tag for `RegExp` constructor", () => {
+      expect(tagFromNativeClass(RegExp)).toBe(NATIVE_TAGS.RegExp);
+    });
+
+    it("returns `null` for unrecognized constructors", () => {
+      expect(tagFromNativeClass(WeakMap)).toBe(null);
+      expect(tagFromNativeClass(Promise)).toBe(null);
+    });
+
+    it("returns `HasToJSON` for class with `toJSON` on prototype", () => {
+      class WithToJSON {
+        toJSON() {
+          return { x: 1 };
+        }
+      }
+      expect(tagFromNativeClass(WithToJSON)).toBe(NATIVE_TAGS.HasToJSON);
+    });
+
+    it("returns `HasToJSON` for subclass inheriting `toJSON`", () => {
+      class Base {
+        toJSON() {
+          return "base";
+        }
+      }
+      class Sub extends Base {}
+      expect(tagFromNativeClass(Sub)).toBe(NATIVE_TAGS.HasToJSON);
+    });
+
+    it("returns `Date` tag for `Date` (not `HasToJSON` despite `Date.prototype.toJSON`)", () => {
+      expect(tagFromNativeClass(Date)).toBe(NATIVE_TAGS.Date);
+    });
+
+    it("returns `null` for class without `toJSON`", () => {
+      class Plain {}
+      expect(tagFromNativeClass(Plain)).toBe(null);
+    });
+  });
+});
